% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAM2VCF.R
\name{BAM2VCF}
\alias{BAM2VCF}
\title{Variant discovery on a BAM file}
\usage{
BAM2VCF(bam, grange, destination)
}
\arguments{
\item{bam}{\code{character} Path to the BAM file}

\item{grange}{\code{GRange} GRange of the BAM file to search for variants}

\item{destination}{\code{character} Path to the VCF file}
}
\description{
Discover variants of a BAM file and write them in a VCF file
}
