% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCountsDS.R
\name{cellCountsDS}
\alias{cellCountsDS}
\title{Estimate cell counts from methylation data}
\usage{
cellCountsDS(x, cellTypeRef = "blood gse35069 complete")
}
\arguments{
\item{x}{an eSet-class object having beta values in the experimental data.}

\item{cellTypeRef}{Character string name of the cell type reference to use for estimating cell counts. 
See meffil.list.cell.type.references() for a list of available references in the package 'meffil'.#'}
}
\value{
a data frame with samples as rows and variables as columns having the cell count estimates
}
\description{
This function uses the function 'meffil.estimate.cell.counts.from.betas' from pacakge 'meffil'
}
\author{
Gonzalez, JR.
}
