% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeN.R
\name{computeN}
\alias{computeN}
\title{Compute N}
\usage{
computeN(gamma)
}
\arguments{
\item{gamma}{\code{numeric} Value of gamma to compute N}
}
\value{
A number to be used as N on re-sampling methods.
}
\description{
Calculates N given gamma.
}
\details{
Computation is performed following Theorem 15 of Rubinstein, Benjamin IP, and Francesco Aldà. "Pain-free random differential privacy with sensitivity sampling." International Conference on Machine Learning. PMLR, 2017.
}
