% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOpalFiles.R
\name{createOpalFiles}
\alias{createOpalFiles}
\title{Creating genetic expression and phenotype dictionary and data files}
\usage{
createOpalFiles(geneExpData, phenoFile, inputDIR = getwd(),
  estimateCellCounts = FALSE, cellTypeRef = "blood gse35069 complete",
  geneExpTableName = "geneExp", phenoTableName = "pheno")
}
\arguments{
\item{geneExpData}{an Expression Set object or name of genetic expression csv/txt file}

\item{phenoFile}{parameter used with csv/txt files to specify name of phenotype file}

\item{inputDIR}{optional parameter used with csv/txt files to specify directory containing data files}

\item{estimateCellCounts}{optional TRUE/FALSE parameter which determines whether epigenomic data is calculated and added to phenotypic data file}

\item{cellTypeRef}{cell type reference used to estimate cell counts}

\item{geneExpTableName}{optional parameter which sets the gene expression table name in Opal}

\item{phenoTableName}{optional parameter which sets the phenotype data table name in Opal}
}
\description{
Creating genetic expression and phenotype dictionary and data files
}
\details{
Function processes an Expression Set object or csv/txt files with
genetic expression and related phenotype data to create xls dictionary
and csv data files used for Opal data table creation in DataShield.
}
\examples{



}
