% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRSEDS.R
\name{createRSEDS}
\alias{createRSEDS}
\title{create a RangedSummarizedExperiment from a RNAseq count table and 
a phenotypes table}
\usage{
createRSEDS(rnaseq, phe, ...)
}
\arguments{
\item{rnaseq}{\code{data.frame} RNAseq count table}

\item{phe}{\code{data.frame} Phenotypes table}

\item{...}{\code{character vector} Columns of the counts table to be used as 
annotation when creating the SummarizedExperiment. If \code{NULL}, biomaRt will be used to find 
annotation using the first column of the counts table (EntrezID)}
}
\value{
A \code{RangedSummarizedExperiment} object
}
\description{
create a RangedSummarizedExperiment from a RNAseq count table and 
a phenotypes table
}
\details{
The RNAseq count table has to have the following structure: \cr
- First column: Entrez identificator (column named "EntrezID") \cr
- Second column: Annotated symbol (optional, but if present must be on this position, if not present, 
make sure to set annot = FALSE) \cr
- All the remaining columns: One col. per individual \cr
The phenotypes table has to have the following structure: \cr
- First column: ID of the individuals. To match with the column names 
of the RNAseq table \cr
- All the remaining columns: One col. per phenotype
}
