% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2DS.R
\name{DESeq2DS}
\alias{DESeq2DS}
\title{Differential expression analysis on the server-side}
\usage{
DESeq2DS(vars, set, test, fitType, sfType, betaPrior, reduced, contrast)
}
\arguments{
\item{vars}{variables for the model}

\item{set}{The data file}

\item{test}{"Wald" or "LRT"}

\item{fitType}{"parametric", "local", or "mean"}

\item{sfType}{"ratio", "poscounts", or "iterate"}

\item{betaPrior}{TRUE or FALSE}

\item{reduced}{test="LRT"}

\item{contrast}{model matrix contrast}
}
\description{
Performs the Differential expression analysis based on the Negative Binomial distribution on the server-side
}
\details{
This function perform a differential expression analysis 
for \code{Count matrix}, \code{htseq-count} and \code{SummarizedExperiment} input
}
\author{
L. Abarrategui for DataSHILED development team
}
