% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplaceNoise.R
\name{dlaplace}
\alias{dlaplace}
\title{Random laplace distribution generator}
\usage{
dlaplace(y, m = 0, s = 1, log = FALSE)
}
\arguments{
\item{y}{\code{numeric} Samples}

\item{m}{\code{integer} (default \code{0}) Mean of Laplace distribution}

\item{s}{\code{integer} (default \code{1}) Standard deviation of Laplace distribution}

\item{log}{\code{bool} (default \code{FALSE}) Apply \code{exp} transformation to the distribution}
}
\value{
Vector of numbers
}
\description{
Extracted from https://rpubs.com/mengxu/gaussian-and-laplace-noise
}
