% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGWASDS.R
\name{.fittedValues_exponential}
\alias{.fittedValues_exponential}
\title{Title}
\usage{
.fittedValues_exponential(mod_values, mod_values_names, x)
}
\arguments{
\item{mod_values}{}

\item{mod_values_names}{}

\item{x}{}

\item{intercept}{}
}
\value{

}
\description{
Title
}
