% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoding.R
\name{.one_hot_encoding}
\alias{.one_hot_encoding}
\title{Create dummy variables}
\usage{
.one_hot_encoding(df, columns = "season")
}
\arguments{
\item{df}{Input data frame}

\item{columns}{Columns to create dummies}
}
\description{
Create dummy variables
}
\details{
Function adapted from https://stackoverflow.com/a/66817515
}
