% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boxplotMethy.R
\name{ds.boxplotMethy}
\alias{ds.boxplotMethy}
\title{Boxplot summary of methylation values split by phenotypic variable levels}
\usage{
ds.boxplotMethy(cpg, phenoVar, molecular.data, pheno.data, datasources)
}
\arguments{
\item{cpg}{name of the CpG site the boxplot will be created for}

\item{phenoVar}{phenotypic variable, each level of which a boxplot will be created for}

\item{molecular.data}{name of the DataSHIELD object to which the methylation data set has been assigned}

\item{pheno.data}{name of the DataSHIELD object to which the phenotype data set has been assigned}

\item{datasources}{Opal object or list of opal objects denoting the opal server(s) information}
}
\description{
Generating boxplot summary of methylation values split by phenotypic variable levels
}
\details{
Function generates a boxplot summary of methylation data by phenotypic condition 
for a specified CpG site. If multiple studies are analysed the boxplots provide representations
of the pooled data. Within a single chart a boxplot is generated for each level of the user 
specified phenotypic variable.
}
\examples{

}
