% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.featureData.R, R/ds.varLabels.R
\name{ds.featureData}
\alias{ds.featureData}
\title{Retrieve information on features from eSets.}
\usage{
ds.featureData(object, connections = NULL)

ds.featureData(object, connections = NULL)
}
\arguments{
\item{object}{name of the DataSHIELD object to which the ExpresionSet has been assigned}

\item{connections}{....}

\item{object}{name of the DataSHIELD object to which the ExpresionSet has been assigned}

\item{connections}{....}
}
\value{
\code{featureNames} returns a (usually long!) character vector uniquely identifying each feature.

a character vector of measured variables.
}
\description{
This function is similar to Bioconductor function \code{featureData}

This function is similar to Bioconductor function \code{varLabels}
}
\examples{


}
