% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRDS.R
\name{edgeRDS}
\alias{edgeRDS}
\title{Differential expression analysis on the server-side}
\usage{
edgeRDS(
  set,
  variable_names,
  intercept,
  dispersion,
  normalization,
  contrast,
  levels,
  test,
  coef
)
}
\arguments{
\item{set}{\code{RangedSummarizedExperiment} object}

\item{variable_names}{grouping variables used to perform differential expression analysis}

\item{intercept}{model intercept}

\item{dispersion}{type of dispersion to estimate}

\item{normalization}{normalization method}

\item{contrast}{character or numeric vector specifying the contrast}

\item{levels}{character or factor vector specifying the names of the contrast parameters}

\item{test}{test for differential expression analysis}

\item{coef}{a character or an integer indicating the coefficient to be tested equal to zero}
}
\description{
Performs a differential expression analysis based on
 the negative binomial distribution.
}
\details{
This function performs a differential expression analysis 
for \code{SummarizedExperiment} input using functions from 
\code{edgeR} Bioconductor's package. 


\code{edgeRDS} is a server-side aggregate function called by \code{ds.edgeR}.
}
\author{
L. Abarrategui for DataSHILED development team
}
