% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGAhtsgetResourceClient.R
\docType{class}
\name{EGAhtsgetResourceClient}
\alias{EGAhtsgetResourceClient}
\title{EGA htsget resource client}
\format{
A R6 object of class EGAhtsgetResourceClient
}
\description{
Build a GDS handler from a resource object describing access to a 
EGA valid credentials, pointing to a VCF or BAM file to be converted into a GDS file.
}
\section{Super classes}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{\link[resourcer:FileResourceClient]{resourcer::FileResourceClient}} -> \code{EGAhtsgetResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EGAhtsgetResourceClient-new}{\code{EGAhtsgetResourceClient$new()}}
\item \href{#method-EGAhtsgetResourceClient-getConnection}{\code{EGAhtsgetResourceClient$getConnection()}}
\item \href{#method-EGAhtsgetResourceClient-getValue}{\code{EGAhtsgetResourceClient$getValue()}}
\item \href{#method-EGAhtsgetResourceClient-close}{\code{EGAhtsgetResourceClient$close()}}
\item \href{#method-EGAhtsgetResourceClient-clone}{\code{EGAhtsgetResourceClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asDataFrame'><code>resourcer::ResourceClient$asDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="FileResourceClient" data-id="downloadFile"><a href='../../resourcer/html/FileResourceClient.html#method-FileResourceClient-downloadFile'><code>resourcer::FileResourceClient$downloadFile()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EGAhtsgetResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-EGAhtsgetResourceClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EGAhtsgetResourceClient$new(resource)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EGAhtsgetResourceClient-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-EGAhtsgetResourceClient-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EGAhtsgetResourceClient$getConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EGAhtsgetResourceClient-getValue"></a>}}
\if{latex}{\out{\hypertarget{method-EGAhtsgetResourceClient-getValue}{}}}
\subsection{Method \code{getValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EGAhtsgetResourceClient$getValue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EGAhtsgetResourceClient-close"></a>}}
\if{latex}{\out{\hypertarget{method-EGAhtsgetResourceClient-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EGAhtsgetResourceClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EGAhtsgetResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EGAhtsgetResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EGAhtsgetResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
