% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGAhtsgetResourceResolver.R
\docType{class}
\name{EGAhtsgetResourceResolver}
\alias{EGAhtsgetResourceResolver}
\title{EGA htsget resource resolver}
\format{
A R6 object of class EGAhtsgetResourceResolver
}
\description{
Build a EGA htsget resource client from a resource object describing access to a 
EGA valid credentials, pointing to a VCF or BAM file to be converted into a GDS file.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{EGAhtsgetResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EGAhtsgetResourceResolver-isFor}{\code{EGAhtsgetResourceResolver$isFor()}}
\item \href{#method-EGAhtsgetResourceResolver-newClient}{\code{EGAhtsgetResourceResolver$newClient()}}
\item \href{#method-EGAhtsgetResourceResolver-clone}{\code{EGAhtsgetResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EGAhtsgetResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-EGAhtsgetResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EGAhtsgetResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EGAhtsgetResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-EGAhtsgetResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EGAhtsgetResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EGAhtsgetResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EGAhtsgetResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EGAhtsgetResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
