% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactHWEDS.R
\name{exactHWEDS}
\alias{exactHWEDS}
\title{Hardy-Weinberg Equilibrium testing}
\usage{
exactHWEDS(
  genoData,
  chromosome,
  geno.counts,
  block.size,
  permute,
  controls,
  controls_column
)
}
\arguments{
\item{genoData}{\code{GenotypeData} object}

\item{chromosome}{\code{character} Chromosome to study. \code{"all"} to study all available chromosomes}

\item{geno.counts}{\code{bool} if \code{TRUE}, genotype counts are returned in the output data.frame}

\item{block.size}{\code{numeric}  number of SNPs to read in at once}

\item{permute}{\code{bool} logical indicator for whether to permute alleles before calculations}

\item{controls}{\code{bool} logical to calculate the HWE test only on the controls}

\item{controls_column}{\code{character} name of the case/controls column of the covariates 
used to create the GenotypeData object. Only used if \code{controls = TRUE}}
}
\value{

}
\description{
This function performs exact Hardy-Weinberg Equilibrium testing (using Fisher's Test) 
over a selection of SNPs. It also counts genotype, calculates allele frequencies, 
and calculates inbreeding coefficients.
}
