% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprsDS.R
\name{exprsDS}
\alias{exprsDS}
\title{Retrieve information on molecular data recorded in eSet and ExpressionSet-derived classes.}
\usage{
exprsDS(x)
}
\arguments{
\item{x}{Object, possibly derived from eSet-class.}
}
\value{
a matrix features as rows, individuals/samples as columns
}
\description{
This function is similar to the Biobase function 'exprs'.
}
\author{
Gonzalez, JR.
}
