% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPhenoFromGDSDS.R
\name{extractPhenoFromGDSDS}
\alias{extractPhenoFromGDSDS}
\title{Extract phenotype table from GenotypeData object}
\usage{
extractPhenoFromGDSDS(gds)
}
\arguments{
\item{gds}{\code{GenotypeData} object}
}
\value{
\code{data.table} of the phenotypes
}
\description{
Extract phenotype table from GenotypeData object
}
