% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastAlleleFrequencyDS.R
\name{fastAlleleFrequencyDS}
\alias{fastAlleleFrequencyDS}
\title{Title}
\usage{
fastAlleleFrequencyDS(genoData, snpBlock)
}
\arguments{
\item{genoData}{}

\item{snpBlock}{}
}
\value{

}
\description{
Title
}
