% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGWASDS.R
\name{fastGWAS_getFitted.values}
\alias{fastGWAS_getFitted.values}
\title{Title}
\usage{
fastGWAS_getFitted.values(x, covars, mod_names, output_family, ...)
}
\arguments{
\item{x}{}

\item{...}{}

\item{y}{}
}
\value{

}
\description{
Title
}
