% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGWASDS.R
\name{fastGWAS_getResiduals}
\alias{fastGWAS_getResiduals}
\title{Title}
\usage{
fastGWAS_getResiduals(x, fitted.values, objective_variable, output_family)
}
\arguments{
\item{x}{}

\item{y}{}

\item{...}{}
}
\value{

}
\description{
Title
}
