% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGWASDS.R
\name{fastGWAS_PHENO_removeNAindiv}
\alias{fastGWAS_PHENO_removeNAindiv}
\title{Remove individuals from the phenotype that have NAs on any of the objective variable or covars}
\usage{
fastGWAS_PHENO_removeNAindiv(pheno, objective, covars)
}
\arguments{
\item{pheno}{\code{PhenoInfoTable} to which remove individuals}

\item{objective}{\code{character} Objective variable}

\item{covars}{\code{character vector} Covariables}
}
\value{
\code{PhenoInfoTable} with individuals trimmed
}
\description{
Remove individuals from the phenotype that have NAs on any of the objective variable or covars
}
