% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGWASDS.R
\name{fastGWAS_S}
\alias{fastGWAS_S}
\title{Extract Geno iterator and Pheno information from GenotypeData GDS containers}
\usage{
fastGWAS_S(genoData, type, snpBlock)
}
\arguments{
\item{genoData}{\code{GenotypeData}}

\item{type}{}

\item{snpBlock}{}
}
\value{

}
\description{
Extract Geno iterator and Pheno information from GenotypeData GDS containers
}
