% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGWASDS.R
\name{fastGWAS_S_means}
\alias{fastGWAS_S_means}
\title{Get mean of the genotype by individual}
\usage{
fastGWAS_S_means(geno, pheno)
}
\arguments{
\item{geno}{\code{GenotypeBlockIterator} Genotype information}

\item{pheno}{\code{PhenoInfoTable} Phenotype information}
}
\value{
\code{named numeric vector} with the means by individual. The names are the individual IDs
}
\description{
Get mean of the genotype by individual
}
