% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureDataDS.R
\name{featureDataDS}
\alias{featureDataDS}
\title{Retrieve information on features recorded in eSet-derived objects.}
\usage{
featureDataDS(x)
}
\arguments{
\item{object}{Object, possibly derived from \code{eSet-class}.}
}
\value{
an object containing information on both variable values and variable meta-data.
}
\description{
This function is similar to the Biobase function \code{featureData}.
}
\author{
Gonzalez, JR.
}
