% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureNamesDS.R
\name{featureNamesDS}
\alias{featureNamesDS}
\title{Retrieve feature and sample names from eSets.}
\usage{
featureNamesDS(x)
}
\arguments{
\item{x}{Object, possibly derived from eSet-class or AnnotatedDataFrame.}
}
\value{
a caracter vector uniquely identifying each feature
}
\description{
This function is similar to the Biobase function 'featureNames'.
}
\author{
Gonzalez, JR.
}
