% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByExprDS.R
\name{filterByExprDS}
\alias{filterByExprDS}
\title{Filter Genes By Expression Level}
\usage{
filterByExprDS(object)
}
\arguments{
\item{object}{\code{eSet}, \code{RangedSummarizedExperiment} on the server}
}
\value{
A \code{RangedSummarizedExperiment} object with filtered genes
}
\description{
Determine which genes have sufficiently large counts to be retained in a statistical analysis. This is a similar function to \code{edgeR::filterByExpr}
}
