% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvarLabelsDS.R
\name{fvarLabelsDS}
\alias{fvarLabelsDS}
\title{Retrieve names of annotation from eSets.}
\usage{
fvarLabelsDS(x)
}
\arguments{
\item{x}{Object, possibly derived from eSet-class or AnnotatedDataFrame.}
}
\value{
a caracter vector uniquely identifying each feature
}
\description{
This function is similar to the Biobase function \code{featureNames}.
}
\author{
Gonzalez, JR.
}
