% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA4GHResourceClient.R
\docType{class}
\name{GA4GHResourceClient}
\alias{GA4GHResourceClient}
\title{GA4GH htsget server resource client}
\format{
A R6 object of class GA4GHResourceClient
}
\description{
Build a GDS handler from a resource object describing access to a 
GA4GH htsget enabled server, pointing to a VCF or BAM file to be converted into a GDS file.
}
\section{Super classes}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{\link[resourcer:FileResourceClient]{resourcer::FileResourceClient}} -> \code{GA4GHResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GA4GHResourceClient-new}{\code{GA4GHResourceClient$new()}}
\item \href{#method-GA4GHResourceClient-getConnection}{\code{GA4GHResourceClient$getConnection()}}
\item \href{#method-GA4GHResourceClient-getValue}{\code{GA4GHResourceClient$getValue()}}
\item \href{#method-GA4GHResourceClient-close}{\code{GA4GHResourceClient$close()}}
\item \href{#method-GA4GHResourceClient-clone}{\code{GA4GHResourceClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asDataFrame'><code>resourcer::ResourceClient$asDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="FileResourceClient" data-id="downloadFile"><a href='../../resourcer/html/FileResourceClient.html#method-FileResourceClient-downloadFile'><code>resourcer::FileResourceClient$downloadFile()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GA4GHResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-GA4GHResourceClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GA4GHResourceClient$new(resource)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GA4GHResourceClient-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-GA4GHResourceClient-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GA4GHResourceClient$getConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GA4GHResourceClient-getValue"></a>}}
\if{latex}{\out{\hypertarget{method-GA4GHResourceClient-getValue}{}}}
\subsection{Method \code{getValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GA4GHResourceClient$getValue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GA4GHResourceClient-close"></a>}}
\if{latex}{\out{\hypertarget{method-GA4GHResourceClient-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GA4GHResourceClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GA4GHResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GA4GHResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GA4GHResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
