% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA4GHResourceResolver.R
\docType{class}
\name{GA4GHResourceResolver}
\alias{GA4GHResourceResolver}
\title{GA4GH htsget server resource resolver}
\format{
A R6 object of class GA4GHResourceResolver
}
\description{
Build a GA4GH resource client from a resource object describing access to a 
GA4GH htsget enabled server, pointing to a VCF or BAM file to be converted into a GDS file.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{GA4GHResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GA4GHResourceResolver-isFor}{\code{GA4GHResourceResolver$isFor()}}
\item \href{#method-GA4GHResourceResolver-newClient}{\code{GA4GHResourceResolver$newClient()}}
\item \href{#method-GA4GHResourceResolver-clone}{\code{GA4GHResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GA4GHResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-GA4GHResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GA4GHResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GA4GHResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-GA4GHResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GA4GHResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GA4GHResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GA4GHResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GA4GHResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
