% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSFileResourceClient.R
\docType{class}
\name{GDSFileResourceClient}
\alias{GDSFileResourceClient}
\title{GDS file resource client}
\format{A R6 object of class GDSFileResourceClient}
\usage{
GDSFileResourceClient
}
\description{
Build a GDS handler from a resource object describing access to a 
GDS file or a VCF file to be converted into a GDS file.
}
\keyword{datasets}
