% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSFileResourceResolver.R
\docType{class}
\name{GDSFileResourceResolver}
\alias{GDSFileResourceResolver}
\title{GDS file resource resolver}
\format{A R6 object of class GDSFileResourceResolver}
\usage{
GDSFileResourceResolver
}
\description{
Build a GDS resource client from a resource object describing access to a 
GDS file or a VCF file to be converted into a GDS file.
}
\keyword{datasets}
