% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsSubset.R
\name{gdsSubset2}
\alias{gdsSubset2}
\title{Title}
\usage{
gdsSubset2(
  parent.gds,
  sub.gds,
  sample.include = NULL,
  snp.include = NULL,
  sub.storage = NULL,
  compress = "LZMA_RA",
  block.size = 5000,
  verbose = TRUE,
  allow.fork = FALSE
)
}
\arguments{
\item{parent.gds}{}

\item{sub.gds}{}

\item{sample.include}{}

\item{snp.include}{}

\item{sub.storage}{}

\item{compress}{}

\item{block.size}{}

\item{verbose}{}

\item{allow.fork}{}
}
\value{

}
\description{
Title
}
\examples{

## ADAPTED FROM GWASTools::gdsSubset
}
