% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCADS.R
\name{geno_pca_pooled_addPC2GenoDS}
\alias{geno_pca_pooled_addPC2GenoDS}
\title{Add PCA results to the phenotype slot}
\usage{
geno_pca_pooled_addPC2GenoDS(geno, pca)
}
\arguments{
\item{geno}{\code{GenotypeData} object}

\item{pca}{\code{data.frame} of the PCA results}
}
\value{
\code{GenotypeData} object
}
\description{
Add the PCA results to be used on an association analysis as covariates
}
