% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCADS.R
\name{geno_pca_pooled_addPCDS}
\alias{geno_pca_pooled_addPCDS}
\title{Add PCA results to geno}
\usage{
geno_pca_pooled_addPCDS(geno, pca, ncomp)
}
\arguments{
\item{pca}{\code{raw} Serialized PCA object}

\item{object}{\code{geno} to add the PCA results}
}
\value{
\code{dt}
}
\description{
Adds the results of the block method to the geno, creating a new dt
}
