% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoDimensionsDS.R
\name{genoDimensionsDS}
\alias{genoDimensionsDS}
\title{Get main dimensions of Genotype data}
\usage{
genoDimensionsDS(x)
}
\arguments{
\item{x}{\code{GenotypeData} or \code{GdsGenotypeReader} object}
}
\value{
\code{list} with results
}
\description{
Get the number of SNPs, number of scans and number of chromosomes on the genotype file
}
