% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeDataDS.R
\name{GenotypeDataDS}
\alias{GenotypeDataDS}
\title{Class GenotypeData.}
\usage{
GenotypeDataDS(
  x,
  covars,
  columnId,
  sexId,
  male_encoding,
  female_encoding,
  case_control_column,
  case,
  control
)
}
\arguments{
\item{x}{a \code{GdsGenotypeReader} object (see GWASTools). It is the object that is
obtained in the opal servers
from a resource of type VCF2GDS}

\item{covars}{a data.frame or a tibble having the metadata of samples (i.e. phenotypes and/or covariates)}

\item{columnId}{\code{numeric} Column of the covars that contains the IDs}

\item{sexId}{\code{numeric} (default \code{NULL}) Column of the covars that contains the sex phenotype}

\item{male_encoding}{\code{character} (default \code{"male"}) String used to encode the male sex
phenotype on the covars table}

\item{female_encoding}{\code{character} (default \code{"female"}) String used to encode the female sex
phenotype on the covars table}

\item{case_control_column}{\code{numeric} (default \code{NULL}) Column that holds the
case/control to relevel to 0/1}

\item{case}{\code{character} (default \code{NULL}) Encoding of the case of the \code{case_control_column}}

\item{control}{\code{character} (default \code{NULL}) Encoding of the control of the \code{case_control_column}}
}
\value{
...
}
\description{
Container for storing genotype data from a GWAS together with the 
metadata associated with the subjects (e.g. phenotypes and covariates) and SNPs 
involved in the study. This is a wrapper of \code{GenotypeData} function from GWASTools package.
}
\author{
Gonzalez, JR.
}
