% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChromosomeNamesDS.R
\name{getChromosomeNamesDS}
\alias{getChromosomeNamesDS}
\title{Get names of chromosomes}
\usage{
getChromosomeNamesDS(genoData)
}
\arguments{
\item{genoData}{\code{GenotypeData} object}
}
\value{
A \code{list} with: \cr
-autosomes: integer codes for the autosomes \cr
-Xchr: integer code for the X chromosome \cr
-pseudoautosomalRegionXY: integer code for the pseudoautosomal region \cr
-Ychr: integer code for the Y chromosome \cr
-mitochondrial: integer code for mitochondrial SNPs \cr
}
\description{
Get the names of the chromosomes of a GenotypeData object
}
