% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPSbyGenDS.R
\name{getSNPSbyGenDS}
\alias{getSNPSbyGenDS}
\title{Subset GDS with gene(s)}
\usage{
getSNPSbyGenDS(gds, old_assign, ...)
}
\arguments{
\item{gds}{\code{GdsGenotypeReader} GDS object}

\item{old_assign}{\code{bool} Internal variable used to re-read an old gds, use at your own risk.}

\item{...}{\code{character} Genes to subset the GDS}
}
\value{
GDS object
}
\description{
Subset GDS with gene(s)
}
