% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariable.R
\name{getVariable}
\alias{getVariable}
\title{Get slot of GDS object}
\usage{
getVariable(x, slot)
}
\arguments{
\item{x}{\code{GenotypeData} or \code{GdsGenotypeReader} object}

\item{slot}{\code{character} slot to retrieve}
}
\value{
Vector of characters
}
\description{
Available slots are limited to: \code{snp.id}, \code{snp.rs.id}, \code{snp.position}, 
\code{snp.chromosome} and \code{snp.allele}
}
