% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSNPDS.R
\name{glmSNPDS}
\alias{glmSNPDS}
\title{Fits a GLM to assess association between the outcome (binary variable) and a given SNP}
\usage{
glmSNPDS(snp, snps, gds, covars, vars, connections, family = "binomial")
}
\value{
a vector with effect estimates, standard error and associated p-value
}
\description{
To be supplied
}
\author{
Gonzalez, JR.
}
