% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASDS.R
\name{GWASDS}
\alias{GWASDS}
\title{Genome-wide association analysis (GWAS) on the server-side}
\usage{
GWASDS(genoData, outcome, covars = NULL, family = "binomial", snpBlock, ...)
}
\arguments{
\item{genoData}{a \code{GenotypeData} object which is a container for storing genotype data
from a GWAS toghether with the metadata associated with the subjects (i.e. phenotypes and/or covariates)
and SNPs}

\item{outcome}{A character string specifying the name of the outcome variable in \code{genoData}}

\item{covars}{A vector of character strings specifying the names of the fixed effect covariates 
in \code{genoData}; an intercept term is automatically included. If NULL (default) the only fixed effect 
covariate is the intercept term}

\item{family}{A description of the generalized linear model used. The defatul is "binomial" that is defined
for case/control studies. Quantitative traits can be analyzed by using "gaussian". Other values are accepted.}

\item{snpBlock}{an integer specifying the number of SNPs in an iteration block. See \code{GenotypeIterator} 
function in GWASTools package.}

\item{...}{other arguments of \code{fitNullModel} function in GENESIS package}
}
\value{
a matrix with SNPs ordered by p-values
}
\description{
Performs GWAS using \code{GENESIS} Bioconductor package
}
\author{
Gonzalez, JR.
}
