% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexGdsnDS.R
\name{indexGdsnDS}
\alias{indexGdsnDS}
\title{Get the specific GDS node}
\usage{
indexGdsnDS(x, ...)
}
\arguments{
\item{x}{an object of class \code{gdsn.class} (a GDS node), or \code{gds.class} (a GDS file)}

\item{...}{other parameters of  \code{index.gdsn} function}
}
\value{
An object of class \code{gdsn.class} for the specified node
}
\description{
This function is similar to function \code{index.gdsn} in the 'gdsfmt' package.
}
\author{
Gonzalez, JR.
}
