% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplaceNoise.R
\name{Laplace_noise_generator}
\alias{Laplace_noise_generator}
\title{Title}
\usage{
Laplace_noise_generator(mu, b, n.noise)
}
\arguments{
\item{mu}{}

\item{b}{}

\item{n.noise}{}
}
\value{

}
\description{
Title
}
