% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDS.R
\name{limmaDS}
\alias{limmaDS}
\title{Differential expression analysis using limma}
\usage{
limmaDS(model, eSet, sva)
}
\value{
a matrix with genes ordered by p-value
}
\description{
To be supplied
}
\author{
Gonzalez, JR.
}
