% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDS2.R
\name{limmaDS2}
\alias{limmaDS2}
\title{Differential expression analysis using limma on the server-side}
\usage{
limmaDS2(Set, res, type, contrasts, coef, annotCols, robust, sort.by)
}
\arguments{
\item{Set}{either a \code{ExpressionSet} or a \code{RangedSummarizedExperiment}}

\item{type}{...}

\item{contrasts}{...}

\item{coef}{...}

\item{annotCols}{variables from the annotation data used in the output}

\item{robust}{Logical value indicating whether robust method is applied in the eBayes function of limma. Default is FALSE.}

\item{variable_names}{grouping variable used to perform differential expression analysis}

\item{covariable_names}{name of variables used in the adjusted models}

\item{levels}{...}

\item{sva}{should differential expression analysis be adjusted by SVA?}

\item{method}{String indicating the method used in the regression (e.g. lmFit function of limma: "ls" or 
"robust". (Default: "ls")}

\item{normalization}{String indicating the normalize method used when using voom for RNAseq data (see normalized.method argument in limma::vomm)}

\item{voomQualityWeights}{Logical value indicating whether limma::voomWithQualityWeights should be used instead of
limma::voom.}

\item{big}{Logical value indicating whether SmartSVA should be used instead of SVA 
(TRUE recommended for methylation or when having large number of samples).}
}
\value{
a matrix with genes ordered by p-value
}
\description{
Performs differential expression analysis using LIMMA
}
\author{
Gonzalez, JR.
}
