% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDS_old.R
\name{limmaDS_old}
\alias{limmaDS_old}
\title{Differential expression analysis using limma}
\usage{
limmaDS_old(model, Set, sva)
}
\value{
a matrix with genes ordered by p-value
}
\description{
To be supplied
}
\author{
Gonzalez, JR.
}
