% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFeatureDS.R
\name{lmFeatureDS}
\alias{lmFeatureDS}
\title{Fits a LM to assess association between the features (outcome) and grouping variable (e.g case/control, condition, ...)}
\usage{
lmFeatureDS(feature, vars, eSet, cellCountsAdjust, connections)
}
\value{
a vector with effect estimates, standard error and associated p-value
}
\description{
To be supplied
}
\author{
Gonzalez, JR.
}
