% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_valid_column_namesDS.R
\name{make_valid_column_namesDS}
\alias{make_valid_column_namesDS}
\title{Rename column names}
\usage{
make_valid_column_namesDS(x)
}
\arguments{
\item{x}{\code{data.frame} Table to change the column names}
}
\value{
\code{data.frame} with the column names updated
}
\description{
Passes the column names through \code{make.names} and rewrites them
}
