% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_valid_namesDS.R
\name{make_valid_namesDS}
\alias{make_valid_namesDS}
\title{Rename column names and character variables}
\usage{
make_valid_namesDS(x)
}
\arguments{
\item{x}{\code{data.frame} Entry table to modify}
}
\value{
\code{data.frame} with the column names and character columns updated
}
\description{
Passes the column names and the character variables through \code{make.names}, this creates 
names without any special characters so there are no problems for accessing any columns or any category 
when using the DataSHIELD parser.
}
