% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFeaturesDS.R
\name{nFeaturesDS}
\alias{nFeaturesDS}
\title{Retrieve number of features from eSets and RangedSummarizedExperiments.}
\usage{
nFeaturesDS(x)
}
\arguments{
\item{x}{Object, possibly derived from eSet-class or RangedSummarizedExperiment.}
}
\value{
a numeric value
}
\description{
This function is similar to the Biobase function \code{dim}.
}
\author{
Gonzalez, JR.
}
