% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSamplesDS.R
\name{nSamplesDS}
\alias{nSamplesDS}
\title{Retrieve number of samples from eSets and RangedSummarizedExperiments.}
\usage{
nSamplesDS(x)
}
\arguments{
\item{x}{Object, possibly derived from eSet-class or RangedSummarizedExperiment.}
}
\value{
a numeric value
}
\description{
This function is similar to the Biobase function \code{dim}.
}
\author{
Gonzalez, JR.
}
