% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opal2FileResourceGetter.R
\docType{class}
\name{Opal2FileResourceGetter}
\alias{Opal2FileResourceGetter}
\title{Opal file resource getter}
\format{
A R6 object of class OpalFileResourceGetter
}
\value{
A OpalFileResourceGetter object.

A logical.

The "resource.file" object.
}
\description{
Creates a new OpalFileResourceGetter instance.

Check that the provided resource has a URL that locates a Opal file: the URL scheme must be "opal+http" or "opal+https" and the path must designate a file web service entry point (i.e. starts with "ws/files/").

Download the file from the Opal file system in a temporary location.
}
\details{
Access a file that is stored in a Opal server. Use Basic authentication header if both
resource's identity and secret are defined, token authentication if secret only is defined.
}
\section{Super class}{
\code{\link[resourcer:FileResourceGetter]{resourcer::FileResourceGetter}} -> \code{Opal2FileResourceGetter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Opal2FileResourceGetter-new}{\code{Opal2FileResourceGetter$new()}}
\item \href{#method-Opal2FileResourceGetter-isFor}{\code{Opal2FileResourceGetter$isFor()}}
\item \href{#method-Opal2FileResourceGetter-downloadFile}{\code{Opal2FileResourceGetter$downloadFile()}}
\item \href{#method-Opal2FileResourceGetter-clone}{\code{Opal2FileResourceGetter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="FileResourceGetter" data-id="extractFileName"><a href='../../resourcer/html/FileResourceGetter.html#method-FileResourceGetter-extractFileName'><code>resourcer::FileResourceGetter$extractFileName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="FileResourceGetter" data-id="makeDownloadDir"><a href='../../resourcer/html/FileResourceGetter.html#method-FileResourceGetter-makeDownloadDir'><code>resourcer::FileResourceGetter$makeDownloadDir()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opal2FileResourceGetter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Opal2FileResourceGetter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opal2FileResourceGetter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opal2FileResourceGetter-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-Opal2FileResourceGetter-isFor}{}}}
\subsection{Method \code{isFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opal2FileResourceGetter$isFor(resource)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opal2FileResourceGetter-downloadFile"></a>}}
\if{latex}{\out{\hypertarget{method-Opal2FileResourceGetter-downloadFile}{}}}
\subsection{Method \code{downloadFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opal2FileResourceGetter$downloadFile(resource, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opal2FileResourceGetter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Opal2FileResourceGetter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opal2FileResourceGetter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
