% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pDataDS.R
\name{pDataDS}
\alias{pDataDS}
\title{Retrieve information on experimental phenotypes recorded in eSet and ExpressionSet-derived classes}
\usage{
pDataDS(x)
}
\arguments{
\item{x}{Object (on the study server), possibly derived from \link{eSet-class} or \link{AnnotatedDataFrame}}
}
\value{
Phenotypes \code{data.frame}
}
\description{
Retrieve information on experimental phenotypes recorded in eSet and ExpressionSet-derived classes
}
