% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkDS.R
\name{plinkDS}
\alias{plinkDS}
\title{Genome-wide association analysis (GWAS) using PLINK}
\usage{
plinkDS(client, ...)
}
\arguments{
\item{client}{...}

\item{...}{}
}
\description{
Performs GWAS using PLINK using shell command lines
}
\author{
Gonzalez, JR.
}
