% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dsMethy.R
\name{plot.dsMethy}
\alias{plot.dsMethy}
\title{qq-plot of p-values output from linear regression analysis of methylation data}
\usage{
\method{plot}{dsMethy}(x, ...)
}
\arguments{
\item{x}{a "ds.Methy" type object containing a column of p-values output from a regression analysis}

\item{...}{All other standard R plot arguments available}
}
\description{
Generating qq-plot of p-values output from linear regression analysis of methylation data
}
\details{
Function produces a qq-plot of the observed vs expected p-values (with confidence bands)
output by the "ds.lmFeature" function. It is an extension of the generic R "plot" function and takes as an 
input "ds.Methy" type objects (one of the classes assigned to the output matrices from "ds.lmFeature")
}
\examples{

}
