% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCADS.R
\name{plotPCADS}
\alias{plotPCADS}
\title{Grouping to plot PCA}
\usage{
plotPCADS(geno, group)
}
\arguments{
\item{group}{}

\item{object}{}
}
\value{

}
\description{
Function used by \code{dsOmicsClient::plotPCA} to get the groupings
}
