% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCASNPSDS.R
\name{plotPCASNPSDS}
\alias{plotPCASNPSDS}
\title{Title}
\usage{
<<<<<<< Updated upstream
plotPCASNPSDS(x, geno, group)
=======
plotPCASNPSDS(x, feno, feno_id, group)
>>>>>>> Stashed changes
}
\arguments{
\item{x}{}

\item{group}{}
}
\value{

}
\description{
Title
}
