% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRSDS.R
\name{PRSDS_aux}
\alias{PRSDS_aux}
\title{Auxiliary function to merge the PRS results to a table by ID}
\usage{
PRSDS_aux(prs_results, prs_results_name, table, id)
}
\description{
Auxiliary function to merge the PRS results to a table by ID
}
