% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGdsnDS.R
\name{readGdsnDS}
\alias{readGdsnDS}
\title{Read data file of a GDS node.}
\usage{
readGdsnDS(x, ...)
}
\arguments{
\item{x}{an object of class \code{gdsn.class}, a GDS node}

\item{...}{other parameters of  \code{read.gdsn} function}
}
\value{
An array, list or data.frame
}
\description{
This function is similar to function \code{read.gdsn} in the 'gdsfmt' package.
}
\author{
Gonzalez, JR.
}
