% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplaceNoise.R
\name{rlaplace}
\alias{rlaplace}
\title{Random laplace generator}
\usage{
rlaplace(n = 1, m = 0, s = 1)
}
\arguments{
\item{n}{\code{integer} (default \code{1}) Number of samples to generate}

\item{m}{\code{integer} (default \code{0}) Mean of Laplace distribution}

\item{s}{\code{integer} (default \code{1}) Standard deviation of Laplace distribution}
}
\value{
Vector of numbers
}
\description{
Extracted from https://rpubs.com/mengxu/gaussian-and-laplace-noise
}
