% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqPreprocDS.R
\name{RNAseqPreprocDS}
\alias{RNAseqPreprocDS}
\title{Filter Genes By Expression Level}
\usage{
RNAseqPreprocDS(object, group)
}
\arguments{
\item{object}{\code{eSet}, \code{RangedSummarizedExperiment} on the server}

\item{group}{vector or factor giving group membership for a oneway layout, if appropriate.}
}
\value{
A \code{RangedSummarizedExperiment} object with filtered genes
}
\description{
Determine which genes have sufficiently large counts to be retained in a statistical analysis. This is a similar function to \code{edgeR::filterByExpr}
}
