% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selSNPDS.R
\name{selSNPDS}
\alias{selSNPDS}
\title{Subset the information of a given SNP in a GDS object.}
\usage{
selSNPDS(genoData, i, strat_variable = NULL, level_number = NULL)
}
\arguments{
\item{genoData}{the GDS object}

\item{i}{the position of the SNP (e.g. column) in the SNP-matrix}
}
\value{
a data frame with samples as rows and selected SNP and variables as columns.
}
\description{
This function creates a data frame with a given SNP and variables from a GDS and data.frame objects.
}
\author{
Gonzalez, JR.
}
