% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpgdsSelectSNPDS.R
\name{snpgdsSelectSNPDS}
\alias{snpgdsSelectSNPDS}
\title{Create a list of candidate SNPs based on specified criteria}
\usage{
snpgdsSelectSNPDS(x, sample.id = NULL, snp.id = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{SNPGDSFileClass}, a SNP GDS file; or characters to specify the file name of SNP GDS}

\item{...}{other parameters of  \code{snpgdsSelectSNPDS} function}
}
\value{
The function returns an integer matrix with values 0, 1, 2 or NA representing the number of reference allele when with.id=FALSE; or list(genotype, sample.id, snp.id) when with.id=TRUE. The orders of sample and SNP IDs in the genotype matrix are actually consistent with sample.id and snp.id in the GDS file, which may not be as the same as the arguments sampel.id and snp.id specified by users.
}
\description{
This function is similar to function \code{snpgdsSelectSNPDS} in the 'SNPRelate' package.
}
\author{
Gonzalez, JR.
}
