% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snptestDS.R
\name{snptestDS}
\alias{snptestDS}
\title{Interface to run SNPtest commands on a ssh connection (with SNPtest 2.5.2 installed)}
\usage{
snptestDS(client, ...)
}
\arguments{
\item{client}{\code{ssh} SSH resource}

\item{...}{\code{character vector} Arguments to pass to the snptest_v2.5.2 software}
}
\value{
List containing: \cr
- Results: Table of results (typical SNPtest output file) \cr
- Output: Console output of the ssh query
}
\description{
Interface to run SNPtest commands on a ssh connection (with SNPtest 2.5.2 installed)
}
