% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetExpressionSet2DS.R
\name{subsetExpressionSet2DS}
\alias{subsetExpressionSet2DS}
\title{Subset ExpressionSet}
\usage{
subsetExpressionSet2DS(eSet, n_ids)
}
\value{
Subseted \code{ExpressionSet}
}
\description{
Subset ExpressionSet
}
