% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGenoDS.R
\name{subsetGenoDS}
\alias{subsetGenoDS}
\title{Subset genotype}
\usage{
subsetGenoDS(genoData, snp_list)
}
\arguments{
\item{genoData}{\code{GenotypeData} or \code{GdsGenotypeReader} object}

\item{snp_list}{\code{character} Use \code{"ethnic_snps"}}
}
\value{
\code{GenotypeData} or \code{GdsGenotypeReader} object
}
\description{
Subset genotype by SNPs linked with ethcity differences from
}
