% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svaDS.R
\name{svaDS}
\alias{svaDS}
\title{Surrogate variable estimates}
\usage{
svaDS(vars, eSet)
}
\arguments{
\item{vars}{a character vector with the variables to be selected.}

\item{x}{an eSet-class object having beta values in the experimental data.}

\item{feature}{a character vector indicanting the selected feature.}
}
\value{
a data frame with samples as rows and selected features and variables as columns.
}
\description{
This function a data frame with surrogate variables used to correct for unwanted variability
}
\author{
Gonzalez, JR.
}
