% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCADS.R
\name{svdPartial}
\alias{svdPartial}
\title{Partial singular value decomposition}
\usage{
svdPartial(x)
}
\arguments{
\item{x}{\code{matrix} Block of data}
}
\value{
\code{matrix} With block step
}
\description{
Block SVD of the SVD block method
}
