% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablePhenoDS.R
\name{tablePhenoDS}
\alias{tablePhenoDS}
\title{Retrieve the number of counts at each factor levels for metadata covariates of eSets and RangedSummarizedExperiments.}
\usage{
tablePhenoDS(x)
}
\arguments{
\item{x}{Object, possibly derived from eSet-class or RangedSummarizedExperiment.}
}
\value{
a numeric value for each factor level
}
\description{
This function is similar to the base function \code{table}.
}
\author{
Gonzalez, JR.
}
