% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhenoDataDS.R
\name{addPhenoDataDS}
\alias{addPhenoDataDS}
\title{Add Phenotype data to ExpressionSet}
\usage{
addPhenoDataDS(x, pheno, identifier)
}
\arguments{
\item{x}{\code{ExpressionSet} ExpressionSet to which add phenotype information}

\item{pheno}{\code{data.frame} Table with the new phenotypes}

\item{identifier}{\code{character} Name of the ID column on the phenotypes data.frame}
}
\value{

}
\description{
Add Phenotype data to ExpressionSet
}
