% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRSEDS.R
\name{createRSEDS}
\alias{createRSEDS}
\title{create a RangedSummarizedExperiment from a RNAseq count table and 
a phenotypes table}
\usage{
createRSEDS(rnaseq, phe)
}
\arguments{
\item{rnaseq}{\code{data.frame} RNAseq count table}

\item{phe}{\code{data.frame} Phenotypes table}
}
\value{
A \code{RangedSummarizedExperiment} object
}
\description{
create a RangedSummarizedExperiment from a RNAseq count table and 
a phenotypes table
}
\details{
The RNAseq count table has to have the following structure: \cr
- First column: Entrez identificator \cr
- Second column: Annotated symbol \cr
- All the remaining columns: One col. per individual \cr
The phenotypes table has to have the following structure: \cr
- First column: ID of the individuals. To match with the column names 
of the RNAseq table \cr
- All the remaining columns: One col. per phenotype
}
