% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRDS.R
\name{edgeRDS}
\alias{edgeRDS}
\title{Differential expression analysis on the server-side}
\usage{
edgeRDS(
  set,
  variable_names,
  intercept,
  dispersion,
  contrast,
  levels,
  test,
  coef
)
}
\description{
Performs the DGE based on
 the negative binomial distribution as a model for count variability,
 including empirical Bayes methods, exact tests, 
 and generalized linear models.
}
\details{
This function perform a DGE analysis 
for \code{SummarizedExperiment} input
}
\author{
L. Abarrategui for DataSHILED development team
}
