% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactHWEDS.R
\name{exactHWEDS}
\alias{exactHWEDS}
\title{Hardy-Weinberg Equilibrium testing}
\usage{
exactHWEDS(
  genoData,
  chromosome,
  geno.counts,
  block.size,
  permute,
  controls_column
)
}
\arguments{
\item{genoData}{\code{GenotypeData} object}

\item{chromosome}{\code{character} Chromosome to study. \code{"all"} to study all available chromosomes}

\item{geno.counts}{\code{bool} if \code{TRUE}, genotype counts are returned in the output data.frame}

\item{block.size}{\code{numeric}  number of SNPs to read in at once}

\item{permute}{\code{bool} logical indicator for whether to permute alleles before calculations}

\item{controls_column}{\code{character} If specified, the control individuals found on this column 
(specified by the encoding \code{1}) will be excluded when calculating the HWE.}
}
\description{
This function performs exact Hardy-Weinberg Equilibrium testing (using Fisher's Test) 
over a selection of SNPs. It also counts genotype, calculates allele frequencies, 
and calculates inbreeding coefficients.
}
