% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactHWEDS.R
\name{exactHWEDS}
\alias{exactHWEDS}
\title{Title}
\usage{
exactHWEDS(genoData, geno.counts, snpStart, snpEnd, block.size, permute)
}
\arguments{
\item{genoData}{}

\item{geno.counts}{}

\item{snpStart}{}

\item{snpEnd}{}

\item{block.size}{}

\item{permute}{}
}
\value{

}
\description{
Title
}
