% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactHWEDS.R
\name{exactHWEDS}
\alias{exactHWEDS}
\title{Hardy-Weinberg Equilibrium testing}
\usage{
exactHWEDS(
  genoData,
  sexcol,
  male,
  female,
  chromosome,
  geno.counts,
  block.size,
  permute
)
}
\arguments{
\item{genoData}{\code{GenotypeData} object}

\item{sexcol}{\code{character} Name of the sex column on the covariates file used to create the 
\code{\link{GenotypeData}} object}

\item{male}{\code{character} Name of the male indicator of the sex column on the covariates file used to create the 
\code{\link{GenotypeData}} object. (Note that it is case sensitive so it's not the same \code{male} than \code{Male})}

\item{female}{\code{character} Name of the female indicator of the sex column on the covariates file used to create the 
\code{\link{GenotypeData}} object. (Note that it is case sensitive so it's not the same \code{female} than \code{Female})}

\item{chromosome}{\code{character} Chromosome to study. \code{"all"} to study all available chromosomes}

\item{geno.counts}{\code{bool} if \code{TRUE}, genotype counts are returned in the output data.frame}

\item{block.size}{\code{numeric}  number of SNPs to read in at once}

\item{permute}{\code{bool} logical indicator for whether to permute alleles before calculations}
}
\value{

}
\description{
This function performs exact Hardy-Weinberg Equilibrium testing (using Fisher's Test) 
over a selection of SNPs. It also counts genotype, calculates allele frequencies, 
and calculates inbreeding coefficients.
}
