% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastAlleleFrequencyDS.R
\name{fastAlleleFrequencyDS}
\alias{fastAlleleFrequencyDS}
\title{Allelic frequency}
\usage{
fastAlleleFrequencyDS(genoData, snpBlock)
}
\arguments{
\item{genoData}{\code{GenotypeData} or \code{GdsGenotypeReader} Genotype data}

\item{snpBlock}{\code{integer} number of SNPs to read at each iteration, 
tune this parameter to improve performance.}
}
\description{
Calculates the frequency of the A allele on a server side GenotypeData object.
}
