% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGWASDS.R
\name{fastGWAS_PHENO_removeNAindiv}
\alias{fastGWAS_PHENO_removeNAindiv}
\title{Title}
\usage{
fastGWAS_PHENO_removeNAindiv(pheno, objective, covars)
}
\arguments{
\item{pheno}{}

\item{objective}{}

\item{covars}{}
}
\value{

}
\description{
Title
}
