% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSFileResourceClient.R
\docType{class}
\name{GDSFileResourceClient}
\alias{GDSFileResourceClient}
\title{GDS file resource client}
\format{
A R6 object of class GDSFileResourceClient
}
\description{
Build a GDS handler from a resource object describing access to a 
GDS file or a VCF file to be converted into a GDS file.
}
\section{Super classes}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{\link[resourcer:FileResourceClient]{resourcer::FileResourceClient}} -> \code{GDSFileResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GDSFileResourceClient-new}{\code{GDSFileResourceClient$new()}}
\item \href{#method-GDSFileResourceClient-getConnection}{\code{GDSFileResourceClient$getConnection()}}
\item \href{#method-GDSFileResourceClient-getValue}{\code{GDSFileResourceClient$getValue()}}
\item \href{#method-GDSFileResourceClient-close}{\code{GDSFileResourceClient$close()}}
\item \href{#method-GDSFileResourceClient-clone}{\code{GDSFileResourceClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asDataFrame'><code>resourcer::ResourceClient$asDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="FileResourceClient" data-id="downloadFile"><a href='../../resourcer/html/FileResourceClient.html#method-FileResourceClient-downloadFile'><code>resourcer::FileResourceClient$downloadFile()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GDSFileResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-GDSFileResourceClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDSFileResourceClient$new(resource)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GDSFileResourceClient-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-GDSFileResourceClient-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDSFileResourceClient$getConnection(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GDSFileResourceClient-getValue"></a>}}
\if{latex}{\out{\hypertarget{method-GDSFileResourceClient-getValue}{}}}
\subsection{Method \code{getValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDSFileResourceClient$getValue(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GDSFileResourceClient-close"></a>}}
\if{latex}{\out{\hypertarget{method-GDSFileResourceClient-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDSFileResourceClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GDSFileResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GDSFileResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDSFileResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
