% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSFileResourceResolver.R
\docType{class}
\name{GDSFileResourceResolver}
\alias{GDSFileResourceResolver}
\title{GDS file resource resolver}
\format{
A R6 object of class GDSFileResourceResolver
}
\description{
Build a GDS resource client from a resource object describing access to a 
GDS file or a VCF file to be converted into a GDS file.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{GDSFileResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GDSFileResourceResolver-isFor}{\code{GDSFileResourceResolver$isFor()}}
\item \href{#method-GDSFileResourceResolver-newClient}{\code{GDSFileResourceResolver$newClient()}}
\item \href{#method-GDSFileResourceResolver-clone}{\code{GDSFileResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GDSFileResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-GDSFileResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDSFileResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GDSFileResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-GDSFileResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDSFileResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GDSFileResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GDSFileResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDSFileResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
