% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsSubset.R
\name{gdsSubset2}
\alias{gdsSubset2}
\title{Write a subset of data in a GDS file to a new GDS file}
\usage{
gdsSubset2(
  parent.gds,
  sub.gds,
  sample.include = NULL,
  snp.include = NULL,
  sub.storage = NULL,
  compress = "LZMA_RA",
  block.size = 5000,
  verbose = TRUE,
  allow.fork = FALSE
)
}
\arguments{
\item{parent.gds}{Name of the parent GDS file}

\item{sub.gds}{Name of the subset GDS file}

\item{sample.include}{Vector of sampleIDs to include in sub.gds}

\item{snp.include}{Vector of snpIDs to include in sub.gds}

\item{sub.storage}{storage type for the subset file; defaults to original storage type}

\item{compress}{The compression level for variables in a GDS file (see \link{GWASTools::add.gdsn} for options)}

\item{block.size}{for GDS files stored with scan,snp dimensions, the number of SNPs to read from the parent 
file at a time. Ignored for snp,scan dimensions.}

\item{verbose}{Logical value specifying whether to show progress information.}

\item{allow.fork}{Logical value specifying whether to enable multiple forks to access the gds file simultaneously.}
}
\description{
Function addapted from GWASTools::gdsSubset. Changed the line 31 (of the source file) to have the argument 
\code{allow.duplicate = TRUE}
}
\author{
Adrienne Stilp
}
