% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeDataDS.R
\name{GenotypeDataDS}
\alias{GenotypeDataDS}
\title{Class GenotypeData.}
\usage{
GenotypeDataDS(x, covars, columnId)
}
\arguments{
\item{x}{..}

\item{covars}{...}

\item{columnId}{...}
}
\value{
...
}
\description{
Container for storing genotype data from a GWAS together with the metadata associated with the subjects (e.g. phenotypes and covariates) and SNPs involved in the study. This is a wrapper of \code{GenotypeData} function from GWASTools package.
}
\author{
Gonzalez, JR.
}
