% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeDataDS.R
\name{GenotypeDataDS}
\alias{GenotypeDataDS}
\title{Class GenotypeData.}
\usage{
GenotypeDataDS(x, covars, ...)
}
\arguments{
\item{x}{an eSet-class object having beta values in the experimental data.}

\item{feature}{a character vector indicanting the selected feature.}

\item{vars}{a character vector with the variables to be selected.}
}
\value{
a data frame with samples as rows and selected features and variables as columns.
}
\description{
Container for storing genotype data from a GWAS together with the metadata associated with the subjects (e.g. phenotypes and covariates) and SNPs involved in the study. This is a wrapper of \code{GenotypeData} function from GWASTools package.
}
\author{
Gonzalez, JR.
}
