% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDS.R
\name{limmaDS}
\alias{limmaDS}
\title{Differential expression analysis using limma on the server-side}
\usage{
limmaDS(
  Set,
  variable_names,
  covariable_names,
  type,
  contrasts,
  levels,
  coef,
  sva,
  annotCols = NULL
)
}
\arguments{
\item{Set}{either a \code{ExpressionSet} or a \code{RangedSummarizedExperiment}}

\item{variable_names}{grouping variable used to perform differential expression analysis}

\item{covariable_names}{name of variables used in the adjusted models}

\item{sva}{should differential expression analysis be adjusted by SVA?}

\item{annotCols}{variables from the annotation data used in the output}
}
\value{
a matrix with genes ordered by p-value
}
\description{
Performs differential expression analysis using LIMMA
}
\author{
Gonzalez, JR.
}
