% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_array_convertDS.R
\name{methylation_array_convertDS}
\alias{methylation_array_convertDS}
\title{Convert methylation arrays}
\usage{
methylation_array_convertDS(eSet, objective_array)
}
\arguments{
\item{eSet}{`ExpressionSet` to convert}

\item{objective_array}{`character` Array to which conver (`"450k"` or `"epic"`)}
}
\value{
`ExpressionSet` converted
}
\description{
Convert methylation arrays
}
