% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSamplesDS.R, R/tablePhenoDS.R
\name{nSamplesDS}
\alias{nSamplesDS}
\title{Retrieve number of samples from eSets and RangedSummarizedExperiments.}
\usage{
nSamplesDS(x)

nSamplesDS(x)
}
\arguments{
\item{x}{Object, possibly derived from eSet-class or RangedSummarizedExperiment.}
}
\value{
a numeric value

a numeric value for each factor level
}
\description{
This function is similar to the Biobase function \code{dim}.

This function is similar to the base function \code{table}.
}
\author{
Gonzalez, JR.

Gonzalez, JR.
}
