% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCADS.R
\name{PCADS}
\alias{PCADS}
\title{Title}
\usage{
PCADS(genoData, pca_rs, pca_means, pca_sd_hw, snpBlock)
}
\arguments{
\item{genoData}{}

\item{rs}{}

\item{means}{}

\item{sd_hw}{}
}
\value{

}
\description{
Title
}
