% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pDataDS.R
\name{pDataDS}
\alias{pDataDS}
\title{Retrieve information on experimental phenotypes recorded in eSet and ExpressionSet-derived classes.}
\usage{
pDataDS(x)
}
\arguments{
\item{x}{Object, possibly derived from eSet-class or AnnotatedDataFrame.}
}
\value{
a data frame with samples as rows, variables as columns
}
\description{
This function is similar to the Biobase function 'pData'.
}
\author{
Gonzalez, JR.
}
