% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCASNPSDS.R
\name{plotPCASNPSDS}
\alias{plotPCASNPSDS}
\title{Grouping to plot PCA}
\usage{
plotPCASNPSDS(x, group)
}
\arguments{
\item{x}{}

\item{group}{}

\item{geno}{}
}
\description{
Function used by \code{dsOmicsClient::plotPCASNPS} to get the groupings
}
