% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutliersDS.R
\name{removeOutliersDS}
\alias{removeOutliersDS}
\title{Filter potential CpG outliers}
\usage{
removeOutliersDS(x)
}
\arguments{
\item{x}{\code{ExpressionSet} or \code{SummarizedExperiment / RangedSummarizedExperiment} object 
to which perform the filtering}
}
\value{

}
\description{
Filter potential CpG outliers
}
