% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetExpressionSetDS.R
\name{subsetExpressionSetDS}
\alias{subsetExpressionSetDS}
\title{Title}
\usage{
subsetExpressionSetDS(eSet, objective_variable, objective_value)
}
\arguments{
\item{objective_value}{}
}
\value{

}
\description{
Title
}
